[implement] [transfer weather ID to weather code, test weather API, works fine]

+++ Drone/app/src/main/java/com/dayton/drone/network/response/model/GetWeatherModel.java
@@ -28,8 +28,8 @@ public class GetWeatherModel {
     }
 
     class Wind {
-        int speed;
-        int deg;
+        float speed;
+        float deg;
     }
 
     class Clouds {
